+++ org.osate.ge/src/org/osate/ge/internal/businessObjectHandlers/ClassifierHandler.java
@@ -54,12 +54,12 @@ import org.osate.ge.internal.services.NamingService;
 import org.osate.ge.internal.ui.dialogs.ClassifierOperationDialog;
 import org.osate.ge.internal.ui.dialogs.DefaultCreateSelectClassifierDialogModel;
 import org.osate.ge.internal.util.AadlClassifierUtil;
-import org.osate.ge.internal.util.ClassifierCreationHelper;
-import org.osate.ge.internal.util.ClassifierOperationExecutor;
 import org.osate.ge.internal.util.ImageHelper;
 import org.osate.ge.internal.util.ScopedEMFIndexRetrieval;
 import org.osate.ge.internal.util.StringUtil;
+import org.osate.ge.internal.util.classifiers.ClassifierCreationHelper;
 import org.osate.ge.internal.util.classifiers.ClassifierOperation;
+import org.osate.ge.internal.util.classifiers.ClassifierOperationExecutor;
 import org.osate.ge.internal.util.classifiers.ClassifierOperationPart;
 import org.osate.ge.internal.util.classifiers.ClassifierOperationPartType;
 import org.osate.ge.query.StandaloneQuery;
