+++ org.osate.ge/src/org/osate/ge/internal/ui/dialogs/DefaultCreateSelectClassifierDialogModel.java
@@ -11,8 +11,8 @@ import org.osate.aadl2.Aadl2Factory;
 import org.osate.aadl2.AadlPackage;
 import org.osate.aadl2.PackageSection;
 import org.osate.ge.internal.services.NamingService;
-import org.osate.ge.internal.util.ClassifierCreationHelper;
 import org.osate.ge.internal.util.ScopedEMFIndexRetrieval;
+import org.osate.ge.internal.util.classifiers.ClassifierCreationHelper;
 import org.osate.ge.internal.util.classifiers.ClassifierOperation;
 import org.osate.ge.internal.util.classifiers.ClassifierOperationPart;
 import org.osate.ge.internal.util.classifiers.ClassifierOperationPartType;
